Fixed AnimatedGIF render texture size
The render texture should be 138x105.  The actual number of LED pixels is less because successive columns are vertically offset.  We use twice the number of LEDs wide and high as the render texture dimensions.  This allows the vertically offset LEDs to land on on existing image coordinates but has some aliasing issues.  For example, for a 2x2 grid of image pixels, two adjacent LEDs represent the bottom left and top right pixels in that grid.