The render target should be twice the naive resolution.
Since every other strip is offset by half vertically, we should render into an image twice the dimensions of the LEDs so that even the offset LEDs will fall on a whole pixel of the render target.  We are currently ignoring half the pixels in the render target.  i.e. imagine a 2x2 square of pixels in the render target.  Two adjacent LEDs where one is offset vertically will effectively sample the bottom left and top right pixels in the corresponding 2x2 pixel grid from the render target.