6" spacing fixture.
Fixture is generated based on dimensions in Dimensions.java.  Currently it just logs a JsonFixture object to the log file which needs to be copied out into it's own .lxf file and placed into lib\Fixtures\myfixture.lxf.  At which point, you should be able to add a new 'myfixture' fixture through the UI.  bvfixture2.lxf is the current best fit generated fixture.  There will be more work to get the outputs properly configured and some minor changes to the numbers of LEDs once the dimensions are perfected.  Also, as you change fixture dimension parameters and generate new fixtures, you also probably need to adjust the size and x and y offsets for the background texture in BG_IMG.java