Added debug frame for verification.
Fixed some bugs.  Here is the latest led layout with counts.  The right side of the peak is rendered relative to the peak of the roof whereas the left side of the peak is rendered relative to the flat roof height.  There is some gap, not sure where that artifact is coming from, maybe due to rendering cylinders with some radius.